requirements.txt
cloudpickle==2.2.1
cycler==0.11.0
Cython==0.29.2
dask==1.0.0
decorator==4.4.2
Django==2.1.5
kiwisolver==1.3.1
mahotas==1.4.5
matplotlib==3.3.4
networkx==2.5.1
numpy==1.16.0
Pillow==5.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.1.1
PyYAML==6.0.1
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.2.0
six==1.16.0
toolz==0.12.0